requirements.txt
python-dateutil>=2.6.0
pytz
requests>=2.17.0
requests-mock>=1.11.0